Moved all zstd helper functions to zstd_helpers in 'debug'
